+++ pytorch_lightning/metrics/classification/__init__.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 from pytorch_lightning.metrics.classification.accuracy import Accuracy
 from pytorch_lightning.metrics.classification.precision_recall import Precision, Recall
-from pytorch_lightning.metrics.classification.f_beta import Fbeta
+from pytorch_lightning.metrics.classification.f_beta import FBeta, F1