+++ pytorch_lightning/metrics/functional/__init__.py
@@ -18,8 +18,6 @@ from pytorch_lightning.metrics.functional.classification import (
     average_precision,
     confusion_matrix,
     dice_score,
-    f1_score,
-    fbeta_score,
     multiclass_precision_recall_curve,
     multiclass_roc,
     precision,
@@ -44,3 +42,4 @@ from pytorch_lightning.metrics.functional.mean_squared_error import mean_squared
 from pytorch_lightning.metrics.functional.mean_squared_log_error import mean_squared_log_error
 from pytorch_lightning.metrics.functional.psnr import psnr
 from pytorch_lightning.metrics.functional.ssim import ssim
+from pytorch_lightning.metrics.functional.f_beta import fbeta, f1